WS2: Router decomposition and docs; StepExecutor dry-run timing
