+++ src/cli_multi_rapid/routing/__init__.py
@@ -1,5 +1,19 @@
-"""Routing utilities including the SimplifiedRouter."""
+"""Routing utilities and Router interfaces."""
 
 from .simplified_router import SimplifiedRouter
+from .router import Router
+from .models import (
+    RoutingDecision,
+    ComplexityAnalysis,
+    ParallelRoutingPlan,
+    AllocationPlan,
+)
 
-__all__ = ["SimplifiedRouter"]
+__all__ = [
+    "SimplifiedRouter",
+    "Router",
+    "RoutingDecision",
+    "ComplexityAnalysis",
+    "ParallelRoutingPlan",
+    "AllocationPlan",
+]
